refactor(multimodal): extract engine-neutral RDMA pixel transport into crates/mm_rdma
Move the NIXL multimodal pixel-transport mechanics out of the gateway
(model_gateway/src/routers/grpc/mm_rdma) into a standalone, engine-neutral
crate `smg-mm-rdma`. Behavior-preserving; the descriptor wire format is
unchanged.

- The crate owns only the NIXL mechanics + wire format and reads no env and
  no router globals. The gateway injects all policy via `RdmaConfig` and
  constructs one exporter (`RdmaExporter::new` / `.export`).
- The optional native dependency (nixl-sys, bindgen-heavy) is isolated behind
  the crate's `nixl` feature; the default build compiles a no-op stub, so
  ordinary gateway builds need no clang/bindgen/libnixl. The gateway's
  `mm-rdma` feature now reads as `["smg-mm-rdma/nixl"]`.
- Gateway owns the exporter singleton + env-derived config
  (multimodal/transport.rs): `mm_rdma_exporter`, `rdma_lane_enabled`,
  `build_rdma_config_from_env`, and the slot-TTL derivation moved here from
  the old module.
- proto_wrapper `try_export_*` call `mm_rdma_exporter().export(slot_key, ..)`;
  `room` renamed to `slot_key` (Rust-internal; wire layout unchanged).
- The pure `SlotPool` (free-list + lease/reclaim + gen framing) is split out
  for hardware-free unit tests (6 tests); the TTL-derivation policy test
  moves to the gateway (transport.rs).
