+++ model_gateway/Cargo.toml
@@ -19,7 +19,7 @@ default = ["grpc-client"]
 grpc-client = []
 grpc-server = []
 opencv-video = ["llm-multimodal/opencv-video"]
-mm-rdma = ["dep:nixl-sys"]
+mm-rdma = ["smg-mm-rdma/nixl"]
 
 vendored-openssl = ["openssl/vendored"]
 
@@ -36,9 +36,6 @@ name = "amg"
 path = "src/main.rs"
 
 [dependencies]
-# Multimodal RDMA pixel transport. Optional because nixl-sys runs bindgen at
-# build time; ordinary gateway builds should not require a NIXL/clang setup.
-nixl-sys = { version = "0.10.1", features = ["stub-api"], optional = true }
 # Workspace dependencies (simple)
 anyhow.workspace = true
 async-trait.workspace = true
@@ -79,6 +76,7 @@ smg-mcp.workspace = true
 kv-index.workspace = true
 smg-data-connector.workspace = true
 llm-multimodal = { workspace = true, default-features = false }
+smg-mm-rdma.workspace = true
 smg-wasm = { workspace = true, features = ["storage-hooks"] }
 smg-mesh.workspace = true
 smg-grpc-client.workspace = true
