+++ model_gateway/src/routers/grpc/mod.rs
@@ -10,7 +10,6 @@ pub(crate) mod common;
 pub(crate) mod context;
 pub(crate) mod epd_encode;
 pub(crate) mod harmony;
-pub(crate) mod mm_rdma;
 pub(crate) mod multimodal;
 pub(crate) mod pd_router; // Used by routers/factory
 pub(crate) mod pipeline;
